implement RTT and RTO and measuring the RTT in the connection phase

+++ src/main/java/org/example/simulator/events/EventGenerator.java
@@ -2,7 +2,7 @@ package org.example.simulator.events;
 
 public abstract class EventGenerator extends Event{
 
-    protected EventGenerator(int delay) {
+    protected EventGenerator(long delay) {
         super(delay);
     }
